tests: separate update suite from cli
